Javascript implementation of the kmeans algorithm with some improvements better suited for getting a color palette.
Javascript implementation of the kmeans algorithm.
This is an improved version that is better suited for getting a color
palette from an image.

Improvements/Changes from the original algorithm include:
I. If a datapoint (color) is too far away from the clusters it will be
added to the set of clusters.
II. With the possible addition of clusters as of I. and a user
specified input 'numberOfClusters' there's a greater dataset to pick
our final palette from
III. With a possible greater dataset as of I. and II. the algorithm
will trim the dataset down to the final palette based on the population
of the clusters.